--- springboot-starter-security-jwt/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.codingapi.springboot.security.configurer.WebSecurityConfigurer,\
-com.codingapi.springboot.security.crypto.MyCryptoConfiguration,\
-com.codingapi.springboot.security.AutoConfiguration
