--- springboot-starter-security-jwt/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,0 @@
-com.codingapi.springboot.security.configurer.WebSecurityConfigurer
-com.codingapi.springboot.security.crypto.MyCryptoConfiguration
-com.codingapi.springboot.security.AutoConfiguration
