--- springboot-starter-security-jwt/src/test/resources/application.properties
@@ -1,29 +1,0 @@
-server.port=8088
-
-codingapi.security.jwt-time=10000
-codingapi.security.jwt-rest-time=5000
-
-# JWT密钥 需大于32位的字符串
-codingapi.security.jwt-secret=codingapi.security.jwt.secretkey
-# JWT AES密钥
-codingapi.security.ase-key=QUNEWCQlXiYqJCNYQ1phc0FDRFgkJV4mKiQjWENaYXM=
-# JWT AES IV
-codingapi.security.aes-iv=QUNYRkdIQEVEUyNYQ1phcw==
-
-# JWT 有效时间(毫秒) 15分钟有效期 1000*60*15=900000
-#codingapi.security.jwt-time=900000
-## JWT 更换令牌时间(毫秒) 10分钟后更换令牌 1000*60*10=600000
-#codingapi.security.jwt-rest-time=600000
-
-# Security 配置 请求权限拦截地址
-codingapi.security.authenticated-urls=/api/**
-# Security 配置 登录地址
-codingapi.security.login-processing-url=/user/login
-# Security 配置 登出地址
-codingapi.security.logout-url=/user/logout
-# Security 配置 不拦截的地址
-codingapi.security.ignore-urls=/open/**
-# 禁用CSRF
-codingapi.security.disable-csrf=true
-# 禁用CORS
-codingapi.security.disable-cors=true
