+++ springboot-starter-security/src/main/java/com/codingapi/springboot/security/filter/SecurityLoginHandler.java
@@ -1,0 +1,14 @@
+package com.codingapi.springboot.security.filter;
+
+import com.codingapi.springboot.security.dto.request.LoginRequest;
+import com.codingapi.springboot.security.gateway.Token;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public interface SecurityLoginHandler {
+
+    void preHandle(HttpServletRequest request, HttpServletResponse response, LoginRequest handler) throws Exception;
+
+    void postHandle(HttpServletRequest request, HttpServletResponse response, LoginRequest handler, Token token);
+
+}
