+++ springboot-starter-security/src/main/java/com/codingapi/springboot/security/gateway/TokenContext.java
@@ -1,0 +1,23 @@
+package com.codingapi.springboot.security.gateway;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+
+public class TokenContext {
+
+    private static final ThreadLocal<String> threadLocal = new ThreadLocal<>();
+
+    private TokenContext() {
+    }
+
+    public static Token current() {
+        return (Token) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    }
+
+    public static void pushExtra(String extra){
+        threadLocal.set(extra);
+    }
+
+    public static String getExtra(){
+        return threadLocal.get();
+    }
+}
