+++ springboot-starter-security/src/main/java/com/codingapi/springboot/security/gateway/TokenGateway.java
@@ -1,0 +1,11 @@
+package com.codingapi.springboot.security.gateway;
+
+import java.util.List;
+
+public interface TokenGateway {
+
+    Token create(String username, String password, List<String> authorities, String extra);
+
+    Token parser(String sign);
+
+}
