+++ springboot-starter-security/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,3 @@
+com.codingapi.springboot.security.configurer.WebSecurityConfigurer
+com.codingapi.springboot.security.jwt.JWTSecurityConfiguration
+com.codingapi.springboot.security.AutoConfiguration
