+++ springboot-starter/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.codingapi.springboot</groupId>
         <artifactId>springboot-parent</artifactId>
-        <version>3.1.10</version>
+        <version>3.2.0.dev</version>
     </parent>
     <artifactId>springboot-starter</artifactId>
